+++ auth/signin.html
@@ -8,7 +8,7 @@
       type="text/css"
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css"
     />
-    <link rel="stylesheet" href="./styles.css" />
+    <link rel="stylesheet" href="../src/styles/auth/styles.css" />
   </head>
   <body>
     <header>
@@ -58,6 +58,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="index.js"></script>
+    <script type="module" src="../src/js/auth/index.js"></script>
   </body>
 </html>
