+++ index.html
@@ -18,8 +18,8 @@
       property="og:description"
       content="세상의 모든 정보를 쉽게 저장하고 관리해 보세요"
     />
-    <link rel="stylesheet" href="./src/reset.css" />
-    <link rel="stylesheet" href="./src/style.css" />
+    <link rel="stylesheet" href="./src/styles/common/reset.css" />
+    <link rel="stylesheet" href="./src/styles/styles.css" />
   </head>
   <body>
     <header>
