+++ src/styles/auth/styles.css
@@ -1,4 +1,4 @@
-@import "../src/global.css";
+@import "../common/global.css";
 
 body {
   padding-top: 23.8rem;
