+++ src/styles/styles.css
@@ -1,4 +1,4 @@
-@import "./global.css";
+@import "./common/global.css";
 
 body {
   display: flex;
